refactor: interface parameters
