--- apps/api-gateway/src/schema/enum/schema.enum.ts
@@ -1,9 +1,0 @@
-export enum SortFields {
-    ID = 'id',
-    CREATED_DATE_TIME = 'createDateTime',
-    NAME = 'name',
-    VERSION = 'version',
-    LEDGER_ID = 'schemaLedgerId',
-    PUBLISHER_DID = 'publisherDid',
-    ISSUER_ID = 'issuerId'
-}
