+++ libs/enum/src/enum.ts
@@ -3,6 +3,16 @@ export enum SortValue {
     DESC = 'desc'
 }
 
+export enum SortFields {
+    ID = 'id',
+    CREATED_DATE_TIME = 'createDateTime',
+    NAME = 'name',
+    VERSION = 'version',
+    LEDGER_ID = 'schemaLedgerId',
+    PUBLISHER_DID = 'publisherDid',
+    ISSUER_ID = 'issuerId'
+}
+
 export enum AgentType {
     AFJ = 'AFJ',
     ACAPY = 'ACAPY'
